refactor: createIcestore
